[重構] getProducts，將拆分函式轉移到 lib/product_tools.js
