+++ controllers/prodCtrller.js
@@ -1,66 +1,11 @@
 const db = require('../models')
-const { Product, Image, Gift, Category } = db
+const { Product, Image, Gift } = db
 
-const Op = require('sequelize').Op
 const moment = require('moment')
 moment.locale('zh-tw')
 
 const PAGE_LIMIT = 30
-const { genQueryString } = require('../lib/tools.js')
-
-function setWhere(req, where) {
-  // 製作 category where
-  let categoryQuery = +req.query.category || 0
-  if (categoryQuery) { where.CategoryId = categoryQuery }
-
-  // 製作 search where (商品名 or 作品名)
-  const searchQuery = req.query.q
-  if (searchQuery) {
-    where[Op.or] = {
-      name: { [Op.like]: `%${searchQuery}%` },
-      '$Series.name$': { [Op.like]: `%${searchQuery}%` }
-    }
-
-    categoryQuery = null  // 搜尋時，關閉分類 active 特效
-  }
-
-  // 製作 tag where (tagQuery 非數字者為特規)
-  let tagQuery = req.query.tag || '所有商品'
-  const isNumber = !isNaN(tagQuery)
-  if (isNumber) {
-    tagQuery = +tagQuery
-    where['$tags.id$'] = tagQuery
-  }
-
-  if (tagQuery === '即將截止預購') {
-    // 從 Date.now 往後取 7 天快截止的商品
-    const date = moment().add(7, 'days').endOf('day')
-
-    where['$tags.id$'] = 1  // 預購中
-    where.deadline = { [Op.lte]: date }
-  }
-
-  return { categoryQuery, searchQuery, tagQuery }
-}
-
-function getPagination(result, PAGE_LIMIT, page) {
-  const totalPages = Math.ceil(result.count / PAGE_LIMIT)
-  const pagesArray = Array.from({ length: totalPages }, (val, index) => index + 1)
-  const prev = (page === 1) ? 1 : page - 1
-  const next = (page === totalPages) ? totalPages : page + 1
-
-  return { pagesArray, prev, next } 
-}
-
-function getBread(categoryQuery, req, res) {
-  if (categoryQuery) {
-    const categories = res.locals.categoryBar
-    const category = categories.find(cate => cate.id === categoryQuery)
-    return category.name
-  }
-
-  if (req.path.includes('search')) return '搜尋商品'
-}
+const { genQueryString, setWhere, getPagination, getBread } = require('../lib/product_tools.js')
 
 module.exports = {
   getProducts: async (req, res) => {
@@ -112,9 +57,10 @@ module.exports = {
       const bread = getBread(categoryQuery, req, res) || '製品一覽'
 
       res.render('products', {
-        js: 'products',
-        css: 'products',
-        getProducts, selectedSort, searchQuery, bread, pagesArray, queryString, categoryQuery, tagQuery, page, prev, next
+        css: 'products', js: 'products', 
+        getProducts, bread,
+        categoryQuery, searchQuery, tagQuery, selectedSort,
+        pagesArray, page, prev, next, queryString
       })
 
     } catch (err) {
